ajoute un test unitaire pour pythonnet

+++ clean_pyd.py
@@ -7,6 +7,7 @@ for root, dirs, files in os.walk("."):
     for f in files : 
         if (".pyd" in f or ".so" in f or ".o" in f or ".def" in f) and "_externals" not in root and "exe.win" not in root :
             filename = os.path.join(root,f)
-            print ("removing ", filename)
-            os.remove(filename)
+            if "pythonnet" not in filename :
+                print ("removing ", filename)
+                os.remove(filename)
